services/config_api/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pyyaml==6.0.1
python-multipart==0.0.6 